feat: Upgrade dependencies
Upgrade all dependencies to the latest major versions and deal with breaking changes in the tool configuration and warnings in the source code.

Replace `chalk` with `picocolors`. It complained about using `import` out of a module. There's still a CJS build output in `storycap`.

Convert build scripts to ESM.

Convert ESLint configuration to flatconfig.